requirements.txt
absl-py==1.2.0
alembic==1.12.1
antlr4-python3-runtime==4.9.3
attrs==23.2.0
autopage==0.5.2
cachetools==4.2.4
cfgv==3.3.1
cliff==3.10.1
cmaes==0.10.0
cmd2==2.4.3
cogdl==0.6
colorlog==6.8.2
cython==0.29.30
decorator==4.4.2
distlib==0.3.8
einops==0.6.1
filelock==3.12.2
flake8==5.0.4
fonttools==4.38.0
fsspec==2023.1.0
gensim==4.2.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grave==0.0.3
greenlet==3.0.3
grpcio==1.60.0
huggingface-hub==0.16.4
identify==2.5.24
importlib-metadata==4.2.0
importlib-resources==5.12.0
joblib==1.3.2
karateclub==1.3.3
kiwisolver==1.4.5
levenshtein==0.23.0
littleutils==0.2.2
llvmlite==0.39.1
mako==1.2.4

markupsafe==2.1.3
mccabe==0.7.0
networkx==2.6.3
ninja==1.11.1.1
nodeenv==1.8.0
numba==0.56.4
oauthlib==3.2.2
ogb==1.3.6
omegaconf==2.3.0
optuna==2.4.0
outdated==0.2.2
packaging==23.2
pandas==1.3.5
pbr==6.0.0
platformdirs==2.6.2
pre-commit==2.21.0
prettytable==3.7.0
protobuf==3.20.0
psutil==5.9.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pyflakes==2.5.0
pygsp==0.5.1
pyparsing==3.1.1
pyperclip==1.8.2
python-levenshtein==0.23.0
python-louvain==0.16
pytorchtools==0.0.2
pytz==2023.3.post1
pyyaml==6.0.1
rapidfuzz==3.4.0
regex==2024.4.16
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.0.2
scipy==1.6.0
sentencepiece==0.2.0
setproctitle==1.3.3
sklearn==0.0
smart-open==7.0.4
sqlalchemy==2.0.29
stevedore==3.5.2
tabulate==0.9.0
tensorboard==2.0.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torchdiffeq==0.2.3
tqdm==4.66.2
transformers==4.30.2
typing-extensions==4.7.1
virtualenv==20.16.2
wcwidth==0.2.13
werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0